+++ monitoring-system/src/main/java/it/uniroma2/sdcc/trafficcontrol/constants/CacheParams.java
@@ -1,0 +1,7 @@
+package it.uniroma2.sdcc.trafficcontrol.constants;
+
+public interface CacheParams {
+
+    String AUTHENTICATION_CACHE_NAME = "auth_hit";
+
+}