+++ monitoring-system/src/main/java/it/uniroma2/sdcc/trafficcontrol/topologies/FirstTopology.java
@@ -1,10 +1,8 @@
 package it.uniroma2.sdcc.trafficcontrol.topologies;
 
-import it.uniroma2.sdcc.trafficcontrol.boltsFirstQuery.FieldsSelectorForRanking;
-import it.uniroma2.sdcc.trafficcontrol.boltsFirstQuery.GlobalRankWindowedBolt;
-import it.uniroma2.sdcc.trafficcontrol.boltsFirstQuery.PartialRankWindowedBolt;
+import it.uniroma2.sdcc.trafficcontrol.boltsFirstQuery.MeanCalculatorBolt;
+import it.uniroma2.sdcc.trafficcontrol.boltsFirstQuery.PartialRankingsBolt;
 import it.uniroma2.sdcc.trafficcontrol.spouts.KafkaSpout;
-import org.apache.storm.topology.base.BaseWindowedBolt;
 import org.apache.storm.tuple.Fields;
 
 import java.util.logging.Logger;
@@ -15,7 +13,7 @@ import static it.uniroma2.sdcc.trafficcontrol.constants.SemaphoreSensorTuple.INT
 import static it.uniroma2.sdcc.trafficcontrol.constants.StormParams.*;
 
 
-public class FirstTopology extends Topology {
+public class FirstTopology extends BaseTopology {
 
     private final static String CLASS_NAME = FirstTopology.class.getSimpleName();
     private final static Logger LOGGER = Logger.getLogger(CLASS_NAME);
@@ -25,7 +23,7 @@ public class FirstTopology extends Topology {
         config.setNumWorkers(NUMBER_WORKERS_SELECTED);
         // Storm default: 1 for workers
         config.setNumAckers(NUMBER_WORKERS_SELECTED);
-        config.setMessageTimeoutSecs(80); // 10 sec in più rispetto alla lunghezza di finestra + interval
+        // config.setMessageTimeoutSecs(80); // 10 sec in più rispetto alla lunghezza di finestra + interval
     }
 
     @Override
@@ -33,30 +31,34 @@ public class FirstTopology extends Topology {
         builder.setSpout(KAFKA_SPOUT, new KafkaSpout(VALIDATED), 2)
                 .setNumTasks(4);
 
-        builder.setBolt(FIELDS_SELECTION_FOR_RANKING, new FieldsSelectorForRanking(), 2)
-                .shuffleGrouping(KAFKA_SPOUT)
+        builder.setBolt(MEAN_CALCULATOR_BOLT, new MeanCalculatorBolt(), 2)
+                .fieldsGrouping(KAFKA_SPOUT, new Fields(INTERSECTION_ID))
                 .setNumTasks(4);
-
-        builder.setBolt(PARTIAL_RANK, new PartialRankWindowedBolt(10).withWindow(
+        /*
+        builder.setBolt(PARTIAL_RANK_BOLT, new PartialRankWindowedBolt(10).withWindow(
                 BaseWindowedBolt.Duration.minutes(1),
                 BaseWindowedBolt.Duration.seconds(10)
-        ), 2)
-                .fieldsGrouping(FIELDS_SELECTION_FOR_RANKING, new Fields(INTERSECTION_ID))
+                ), 2)
+                .fieldsGrouping(MEAN_CALCULATOR_BOLT, new Fields(INTERSECTION_ID))
                 .setNumTasks(4);
-        /*builder.setBolt(PARTIAL_RANK, new PartialRankBolt(10), 2)
-                .fieldsGrouping(FIELDS_SELECTION_FOR_RANKING, new Fields(INTERSECTION_ID))
+        builder.setBolt(PARTIAL_RANK_BOLT, new PartialRankBolt(10), 2)
+                .fieldsGrouping(MEAN_CALCULATOR_BOLT, new Fields(INTERSECTION_ID))
                 .setNumTasks(4);*/
-
-        builder.setBolt(GLOBAL_RANK, new GlobalRankWindowedBolt(10).withWindow(
+        builder.setBolt(PARTIAL_RANK_BOLT, new PartialRankingsBolt(10, 1), 1)
+                .fieldsGrouping(MEAN_CALCULATOR_BOLT, new Fields(INTERSECTION_ID))
+                .setNumTasks(1);
+        /*
+        builder.setBolt(GLOBAL_RANK_BOLT, new GlobalRankWindowedBolt(10).withWindow(
                 BaseWindowedBolt.Duration.minutes(1),
                 BaseWindowedBolt.Duration.seconds(10)
-        ))
-                .globalGrouping(PARTIAL_RANK);
+                ))
+                .globalGrouping(PARTIAL_RANK_BOLT);
+                */
     }
 
     @Override
     public String getClassName() {
-        return "DioCane";
+        return CLASS_NAME;
     }
 
     @Override