+++ monitoring-system/src/main/java/it/uniroma2/sdcc/trafficcontrol/utils/ranking/Rankable.java
@@ -1,4 +1,0 @@
-package it.uniroma2.sdcc.trafficcontrol.utils.ranking;
-
-public interface Rankable {
-}